requirements.txt
django
psycopg2
python-dateutil
social-auth-app-django